Move performance tests into dedicated package
