+++ test/com/intellij/advancedExpressionFolding/performance/PerformanceTest.kt
@@ -1,5 +1,6 @@
-package com.intellij.advancedExpressionFolding
+package com.intellij.advancedExpressionFolding.performance
 
+import com.intellij.advancedExpressionFolding.FoldingTest
 import com.intellij.platform.testFramework.core.FileComparisonFailedError
 import com.intellij.testFramework.PlatformTestUtil
 import org.junit.jupiter.api.AfterAll
@@ -54,8 +55,9 @@ class PerformanceTest : FoldingTest() {
                 }
             }
             val content = """
-package com.intellij.advancedExpressionFolding
+package com.intellij.advancedExpressionFolding.performance
 
+import com.intellij.advancedExpressionFolding.FoldingTest
 import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable
 import kotlin.reflect.KFunction0
 
@@ -80,7 +82,7 @@ $results
 
 }
 """.trimIndent()
-            File("test/com/intellij/advancedExpressionFolding/performanceResult.kt").writeText(content)
+            File("test/com/intellij/advancedExpressionFolding/performance/performanceResult.kt").writeText(content)
         }
     }
